Refactor frontend routing with React Router
